Player Rig Transform on Train now works
The Player can now move on the train and teleport back to the ground without any issues.

The approach of making the player rig a child of the corresponding train part was NOT viable. It caused Fusion to be confused, and after a few ticks, it crashed the game or the character was completely glitching out.

How I solved it:
I realized that the actual Hardware Rig, which is always on the scene, is solely responsible for moving the player. This means that the scripts attached to this object determine the behavior and not the actual network rig that is spawned as a prefab. The spawned prefab just follows the hardware rig, that is already implemented like this by Photon themselves.
I adjusted the "HardwareRig" script to detect when the player object is colliding with a collider on the train with the tag "OnTrain". When that happens, the position is constantly updated to match the position of the object it is colliding with.
I used a similar approach to the method that was already in the "HardwareRig" script for Teleporting the Rig. But that method didn't allow me to teleport around inside the train cabin anymore. So I adjusted it to enable the player to still be able to teleport around but also be moved with the train at the same time.

I already tested it with two players on my PC and it works without any issues.
So our project is hopefully saved!